requirements.txt
# Connexion à SQL Server
pyodbc
SQLAlchemy

# Manipulation avancée des DataFrames
pandas
polars

# Gestion des fichiers CSV et Parquet avec fsspec et adlfs
azure-storage-blob
azure-identity
fsspec
adlfs

Pillow

# Formats de données
pyarrow

# Journalisation et suivi des logs
loguru

# Gestion des variables d'environnement
python-dotenv

# Autres dépendances
numpy
